+++ src/routes/index.tsx
@@ -59,6 +59,10 @@ export default function Router() {
             { path: "createstudenttype", element: <CreateStudentType /> },
           ],
         },
+        {
+          path: "semester",
+          children: [{ path: "create", element: <CreateSemester /> }],
+        },
       ],
     },
     // { path: '*', element: <Navigate to="/404" replace /> }
@@ -88,6 +92,9 @@ const StudentType = Loadable(
 );
 const CreateStudentType = Loadable(
   lazy(() => import("../pages/Settings/CreateStudentType"))
+);
+const CreateSemester = Loadable(
+  lazy(() => import("../pages/Semester/CreateSemester"))
 );
 // Main
 // const ComingSoon = Loadable(lazy(() => import('../pages/ComingSoon')));